Move GUI to its own file
